+++ src/main/java/frc/robot/Constants.java
@@ -23,7 +23,7 @@ public final class Constants {
   }
 
   public final class ControllerConstants {
-    public static final int CONTROLLER_PORT = 1;
+    public static final int CONTROLLER_PORT = 0;
 
     public static final double JOYSTICK_DEADBAND = 0.05;
   }
