+++ src/main/java/frc/robot/commands/DrivebaseTeleop.java
@@ -1,0 +1,45 @@
+// Copyright (c) FIRST and other WPILib contributors.
+// Open Source Software; you can modify and/or share it under the terms of
+// the WPILib BSD license file in the root directory of this project.
+
+package frc.robot.commands;
+
+import java.util.function.Supplier;
+
+import edu.wpi.first.wpilibj2.command.CommandBase;
+
+import frc.robot.subsystems.Drivebase;
+
+public class DrivebaseTeleop extends CommandBase {
+  private final Drivebase drivebase;
+  private final Supplier<Double> m_xaxisSpeedSupplier;
+  private final Supplier<Double> m_zaxisRotateSupplier;
+  private final Supplier<Double> m_throttle;
+
+  public DrivebaseTeleop(
+      Drivebase drivetrain,
+      Supplier<Double> xaxisSpeedSupplier,
+      Supplier<Double> zaxisRotateSupplier,
+      Supplier<Double> throttle) {
+    drivebase = drivetrain;
+    m_xaxisSpeedSupplier = xaxisSpeedSupplier;
+    m_zaxisRotateSupplier = zaxisRotateSupplier;
+    m_throttle = throttle;
+    
+    addRequirements(drivetrain);
+  }
+
+  @Override
+  public void initialize() {}
+
+  @Override
+  public void execute() {}
+
+  @Override
+  public void end(boolean interrupted) {}
+
+  @Override
+  public boolean isFinished() {
+    return false;
+  }
+}
