--- src/main/java/frc/robot/commands/Shoot.java
@@ -1,27 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package frc.robot.commands;
-
-import edu.wpi.first.wpilibj2.command.CommandBase;
-
-public class Shoot extends CommandBase {
-  public Shoot() {
-    // Use addRequirements() here to declare subsystem dependencies.
-  }
-
-  @Override
-  public void initialize() {}
-
-  @Override
-  public void execute() {}
-
-  @Override
-  public void end(boolean interrupted) {}
-
-  @Override
-  public boolean isFinished() {
-    return false;
-  }
-}
